MAINT: Towards unification of all convergence setups with ABC. Anisotropy ok
